0.0.2

+++ csmall-product/src/main/java/com/tqq/csmall/product/services/impl/AttributeTemplateServiceImpl.java
@@ -3,7 +3,10 @@ package com.tqq.csmall.product.services.impl;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.tqq.csmall.product.ex.ServiceException;
 import com.tqq.csmall.product.mapper.AttributeTemplateMapper;
-import com.tqq.csmall.product.pojo.entity.AttributeTemplate;
+import com.tqq.csmall.product.mapper.CategoryAttributeTemplateMapper;
+import com.tqq.csmall.product.mapper.SKUMapper;
+import com.tqq.csmall.product.mapper.SPUMapper;
+import com.tqq.csmall.product.pojo.entity.*;
 import com.tqq.csmall.product.pojo.param.AttributeTemplateAddNewParam;
 import com.tqq.csmall.product.pojo.param.AttributeTemplateUpdateInfoParam;
 import com.tqq.csmall.product.services.IAttributeTemplateService;
@@ -21,6 +24,12 @@ public class AttributeTemplateServiceImpl implements IAttributeTemplateService {
 
     @Autowired
     AttributeTemplateMapper attributeTemplateMapper;
+    @Autowired
+    CategoryAttributeTemplateMapper categoryAttributeTemplateMapper;
+    @Autowired
+    SKUMapper skuMapper;
+    @Autowired
+    SPUMapper spuMapper;
     @Override
     public void addNew(AttributeTemplateAddNewParam attributeTemplateAddNewParam) {
         log.debug("开始处理【添加属性模板】的业务，参数：{}", attributeTemplateAddNewParam);
@@ -57,6 +66,38 @@ public class AttributeTemplateServiceImpl implements IAttributeTemplateService {
             String message = "删除属性模板失败,属性模板信息不存在！";
             log.warn(message);
             throw new ServiceException(ServiceCode.ERR_NOTFOUND,message);
+        }
+
+        /*检测属性模板是否关联了分类属性模板*/
+        QueryWrapper<CategoryAttributeTemplate> queryWrapper1 = new QueryWrapper<>();
+        queryWrapper1.eq("attribute_template_id",id);
+        int countByCategoryAttributeTemplateId = categoryAttributeTemplateMapper.selectCount(queryWrapper1);
+        log.debug("根据属性模板ID统计匹配的相册数量，结果：{}", countByCategoryAttributeTemplateId);
+        if (countByCategoryAttributeTemplateId>0){
+            String message = "删除属性模板不予以执行，属性模板中有分类属性模板未处理";
+            log.warn(message);
+            throw new ServiceException(ServiceCode.ERR_CONFLICT,message);
+        }
+
+        /*spu,sku中也有字段关联到了属性模板，需要检测一下对应id的sku和spu是否存在*/
+        QueryWrapper<SPU> queryWrapper2 = new QueryWrapper<>();
+        queryWrapper2.eq("attribute_template_id",id);
+        int countBySPUId = spuMapper.selectCount(queryWrapper2);
+        log.debug("根据属性模板ID统计匹配的相册数量，结果：{}", countBySPUId);
+        if (countBySPUId>0){
+            String message = "删除属性模板不予以执行，SPU中有关联数据未处理";
+            log.warn(message);
+            throw new ServiceException(ServiceCode.ERR_CONFLICT,message);
+        }
+
+        QueryWrapper<SKU> queryWrapper3 = new QueryWrapper<>();
+        queryWrapper3.eq("attribute_template_id",id);
+        int countBySKUId = skuMapper.selectCount(queryWrapper3);
+        log.debug("根据属性模板ID统计匹配的相册数量，结果：{}", countBySKUId);
+        if (countBySKUId>0){
+            String message = "删除属性模板不予以执行，SKU中有关联数据未处理";
+            log.warn(message);
+            throw new ServiceException(ServiceCode.ERR_CONFLICT,message);
         }
 
         attributeTemplateMapper.deleteById(id);
